requirements.txt
django
djangorestframework
django-filter
mysqlclient
stripe
pillow
django_countries